Move the login form to a separate component
